third_party/blink/web_tests/fast/css/sibling-selectors-dynamic-expected.txt

PASS getComputedStyle(document.getElementById('last')).backgroundColor is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE
